LayoutTests/imported/w3c/web-platform-tests/cookiestore/cookieStore_subscribe_arguments.https.any-expected.txt


FAIL cookieStore.subscribe without name or url in options assert_equals: expected (string) "cookie-name" but got (undefined) undefined
PASS cookieStore.subscribe with empty option
PASS cookieStore.subscribe with invalid url path in option
PASS cookieStore.subscribe is idempotent
PASS CookieStore.unsubscribe is idempotent
